Remove memmove(), memcpy() and memset() completely
Issue #22 in circle-stdlib revealed that the mix of multiple
memcpy() functions in circle-newlib and Circle causes problems,
as the assembler implementation of memmove() for AArch64
unexpectedly called the memcpy() implementation in Circle.

To clean this up once and for all, the implementations for
memmove(), memcpy() and memset() are disabled completely in
circle-newlib, so that only the implementations from Circle
are used.