Adapt to new include strategy
The circle-newlib build now defines the include directory order so that
the Circle include directories appear as system include directories at
the end of the include file search path.

The intention is that the newlib build uses the newlib headers where
Circle has the same headers. The concrete problem was that newlib uses
internal newlib functions from assert.h that were not available when
Circle's assert.h was included.

But there's the reverse problem that Circle's header use internal
macros that do not appear in the newlib headers. Again the assert.h
header is the problem where Circle defines the ASSERT_STATIC() macro.
For this a local header was created that duplicates the necessary
Circle macros.